// lib/trans-short-link/types/Menu.d.ts
export type Menu = {
  id: number;
  name: string;
  isBeta: boolean;
  isNew: boolean;
  isHot: boolean;
  isExpand: boolean;
  key: string;
  fullKey: string;
  deleted: number;
  level: number;
  lastOperator: string | null;
  parentId: number;
  path: string;
  position: number;
  quoteMdId: number;
  subname: string | null;
  tag: string;
  type: string;
  createTime: Date;
  updateTime: Date;
  children?: Menu[];
  showPlatform: number;
  fullPName: string;
  dataModulename: string;
  clickType: string;
  clickTypeLink: string;
  routeKey: string;
  rootRouteKey: string;
  docsRouteKey: string;
};